Implemented Custom hook
