+++ src/components/UseFetch.jsx
@@ -1,0 +1,23 @@
+/* eslint-disable react-hooks/exhaustive-deps */
+import { useState, useEffect } from "react";
+import axios from "axios";
+
+const useFetch = (url) => {
+  const [data, setData] = useState([]);
+  const getNews = async () => {
+    try {
+      const response = await axios.get(url);
+      setData(response?.data?.articles);
+    } catch (error) {
+      console.error("Error fetching news:", error);
+    }
+  };
+
+  useEffect(() => {
+    getNews();
+  }, []);
+
+  return { data };
+};
+
+export default useFetch;
